requirements.txt
Flask>=1.1.0,<2.0.0
hiredis>=1.0.0,<2.0.0
raven>=6.0.0,<7.0.0
rq>=1.1.0,<1.2.0
rq-dashboard>=0.5.0,<0.6.0
simplejson>=3.0.0,<4.0.0
